:recycle: Rename to TextFieldComponent
